+++ v1/plugins/rest/auth.go
@@ -5,6 +5,7 @@
 package rest
 
 import (
+	"cmp"
 	"context"
 	"crypto/rand"
 	"crypto/sha256"
@@ -32,6 +33,7 @@ import (
 	"github.com/lestrrat-go/jwx/v3/jws"
 	"github.com/open-policy-agent/opa/internal/providers/aws"
 	"github.com/open-policy-agent/opa/internal/uuid"
+	"github.com/open-policy-agent/opa/v1/config"
 	"github.com/open-policy-agent/opa/v1/keys"
 	"github.com/open-policy-agent/opa/v1/logging"
 )
@@ -257,6 +259,8 @@ type oauth2ClientCredentialsAuthPlugin struct {
 	signingKeyParsed any
 	tokenCache       *oauth2Token
 	tlsSkipVerify    bool
+	minTLSVersion    uint16
+	cipherSuites     *[]uint16
 	logger           logging.Logger
 }
 
@@ -498,8 +502,10 @@ func (ap *oauth2ClientCredentialsAuthPlugin) NewClient(c Config) (*http.Client, 
 		}
 	}
 
-	// Inherit skip verify from the "parent" settings. Should this be configurable on the credentials too?
+	// Inherit TLS settings from the "parent" config
 	ap.tlsSkipVerify = c.AllowInsecureTLS
+	ap.minTLSVersion = c.minTLSVersion
+	ap.cipherSuites = c.cipherSuites
 
 	ap.logger = c.logger
 
@@ -655,10 +661,15 @@ func (ap *oauth2ClientCredentialsAuthPlugin) requestToken(ctx context.Context) (
 		r.Header.Add(k, v)
 	}
 
-	client := DefaultRoundTripperClient(&tls.Config{
-		MinVersion:         defaultMinTLSVersion,
+	tlsConfig := &tls.Config{
+		MinVersion:         cmp.Or(ap.minTLSVersion, uint16(config.DefaultMinTLSVersion)),
 		InsecureSkipVerify: ap.tlsSkipVerify,
-	}, 10)
+	}
+	if ap.cipherSuites != nil {
+		tlsConfig.CipherSuites = *ap.cipherSuites
+	}
+
+	client := DefaultRoundTripperClient(tlsConfig, 10)
 	response, err := client.Do(r)
 	if err != nil {
 		return nil, err